#3614, further consolidation of biz logic.  need to finish API part
